backend/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
Jinja2==3.1.2
Werkzeug==2.3.7
itsdangerous==2.1.2
flask-mail==0.9.1